A website to classify fish species based on an individual fish's weight and dimensions data.
This website makes use of XGB classifier with Optuna optimizer model for classification fish species. It uses Flask for connecting the model in python
to the website whose backend is in Javascript.

An error encountered while making this, was the 'port difference' error between the Flask API and the web browser, which was solved through the use of Flask COR module.